requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
coverage==7.3.2
    # via -r requirements.in
django==4.2.6
    # via
    #   -r requirements.in
    #   django-crispy-forms
    #   django-htmx
    #   django-mysql
django-crispy-forms==2.1
    # via -r requirements.in
django-htmx==1.17.0
    # via -r requirements.in
django-mysql==4.12.0
    # via -r requirements.in
mysqlclient==2.2.0
    # via -r requirements.in
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
